all components in such named folder
