+++ web/src/client.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { hydrate } from 'react-dom'
 import { BrowserRouter } from 'react-router-dom'
 
-import App from './App'
+import App from 'components/App'
 
 hydrate(
   <BrowserRouter>
