+++ web/src/components/App/index.test.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { render, screen } from '@testing-library/react'
 import '@testing-library/jest-dom/extend-expect'
 
-import App from './App'
+import App from 'components/App'
 import { MemoryRouter } from 'react-router-dom'
 
 describe('<App />', () => {
