+++ web/src/server.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { renderToString } from 'react-dom/server'
 import { StaticRouter } from 'react-router-dom'
 
-import App from './App'
+import App from 'components/App'
 
 let assets: any
